Move sample hit target tests out of widget module
